requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.3
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
click==8.1.7
cryptography==42.0.7
distlib==0.3.8
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
fastapi==0.110.0
filelock==3.14.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
identify==2.5.36
idna==3.6
iniconfig==2.0.0
motor==3.3.2
nodeenv==1.8.0
packaging==24.0
passlib==1.7.4
platformdirs==4.2.1
pluggy==1.4.0
pyasn1==0.6.0
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
pymongo==4.6.2
pytest==8.1.1
pytest-asyncio==0.23.6
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
rsa==4.9
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.28.0
virtualenv==20.26.1
